requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.3
openai==1.14.0

chromadb==0.4.24
sentence-transformers==2.5.1

requests==2.31.0
aiohttp==3.9.3
arxiv==2.1.0

fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3

python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4

tqdm==4.66.2
python-json-logger==2.0.7
colorlog==6.8.2

pytest==8.1.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
rich
streamlit
plotly
matplotlib
seaborn
reportlab==4.0.7